+++ simple/PklProject
@@ -1,0 +1,5 @@
+amends "pkl:Project"
+
+dependencies {
+  ["go"] { uri = "package://pkg.pkl-lang.org/pkl-go/pkl.golang@0.11.0" }
+}
