+++ simple/pkl/AppConfig.pkl
@@ -20,7 +20,8 @@
 @go.Package { name = "github.com/apple/pkl-go-examples/simple/gen/appconfig" }
 module org.pkl.golang.example.AppConfig
 
-import "package://pkg.pkl-lang.org/pkl-go/pkl.golang@0.9.0#/go.pkl"
+import "@go/go.pkl"
+
 import "RedisConfig.pkl"
 
 typealias Port = UInt16(this > 0)
